python-common/cryptotools: move internal crypto caller to new file
